+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased (v0.11.3)
 * Add sample-encode support for computing both XPSNR **and** VMAF using `--xpsnr --and-vmaf`.
+* Add error & hint when mistakenly using `--vmaf NUMBER` without using `--min-vmaf` arg.
 
 # v0.11.2
 * Add arg `--xpsnr-pix-format` for setting an explicit pixel format to use in XPSNR analysis.